Move 2D-only navigation code to its own subfolder
