+++ modules/navigation/2d/godot_navigation_server_2d.h
@@ -31,11 +31,11 @@
 #ifndef GODOT_NAVIGATION_SERVER_2D_H
 #define GODOT_NAVIGATION_SERVER_2D_H
 
-#include "nav_agent.h"
-#include "nav_link.h"
-#include "nav_map.h"
-#include "nav_obstacle.h"
-#include "nav_region.h"
+#include "../nav_agent.h"
+#include "../nav_link.h"
+#include "../nav_map.h"
+#include "../nav_obstacle.h"
+#include "../nav_region.h"
 
 #include "servers/navigation_server_2d.h"
